:recycle: [#46] refactor: Zaaktype create endpoint to reuse code
