+++ backend/src/openbeheer/zaaktype/api/views.py
@@ -42,11 +42,8 @@ from openbeheer.types.ztc import (
     ZaakType,
     ZaakTypeRequest,
 )
-from openbeheer.utils.decorators import handle_service_errors
 from openbeheer.zaaktype.constants import ZAAKTYPE_FIELDSETS
-from rest_framework import status
 from rest_framework.request import Request
-from rest_framework.response import Response
 
 if TYPE_CHECKING:
     from ape_pie import APIClient
@@ -105,7 +102,7 @@ class ZaakTypeSummary(VersionedResourceSummary, kw_only=True, rename="camel"):
     ),
 )
 class ZaakTypeListView(ListView[ZaaktypenGetParametersQuery, ZaakTypeSummary]):
-    data_type = ZaakTypeSummary
+    return_data_type = ZaakTypeSummary
     query_type = ZaaktypenGetParametersQuery
     endpoint_path = "zaaktypen"
 
@@ -128,26 +125,6 @@ class ZaakTypeListView(ListView[ZaaktypenGetParametersQuery, ZaakTypeSummary]):
         if params.catalogus:
             params.catalogus = f"{api_client.base_url}catalogussen/{params.catalogus}"
         return params
-
-    @handle_service_errors
-    def post(self, request: Request, slug: str = "") -> Response:
-        with ztc_client(slug) as client:
-            response = client.post(self.endpoint_path, json=request.data)
-
-        if not response.ok:
-            error = decode(response.content, type=ZGWError)
-            return Response(
-                error,
-                status=response.status_code,
-            )
-
-        data = decode(
-            response.content,
-            type=ZaakType,
-            strict=False,
-        )
-
-        return Response(data, status.HTTP_201_CREATED)
 
 
 ExpandableZaakType = make_expandable(
